% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENull.R
\name{eNull.CM}
\alias{eNull.CM}
\title{Estimate Empirical Null by Central Matching}
\usage{
eNull.CM(z, f, L, U)
}
\arguments{
\item{z}{Observations.}

\item{f}{Fitted mixture density. See \code{\link{Fit.EF}}.}

\item{L}{Lower limit of null neighborhood.}

\item{U}{Upper limit of null neighborhood.}
}
\value{
Estimated mean and variance of the null component, and the proportion
  of observations arising from the null component.
}
\description{
Estimates parameters of the empirical null distribution by central matching
within a "null neighborhood". Observations in the null neighborhood are
assumed to arise from the null component of the mixture distribution.
}
