% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENull.R
\name{eNull.TN}
\alias{eNull.TN}
\title{Maximum Likelihood Estimation of the Empirical Null}
\usage{
eNull.TN(z, L, U, maxit = 10)
}
\arguments{
\item{z}{Observations.}

\item{L}{Lower limit of null neighborhood.}

\item{U}{Upper limit of null neighborhood.}

\item{maxit}{Maximum number of NR iterations.}
}
\value{
Estimated mean and variance of the null component, and the proportion
  of observations arising from the null component.
}
\description{
Estimates parameters of the empirical null distribution by fitting
a truncated normal distribution within a "null neighborhood". Observations
in the null neighborhood are assumed to arise from the null component
of the mixture distribution.
}
