% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENull.R
\name{ENullCM}
\alias{ENullCM}
\title{Estimate Empirical Null by Central Matching}
\usage{
ENullCM(z, f, lower_bound, upper_bound)
}
\arguments{
\item{z}{Observations.}

\item{f}{Fitted mixture density. See \code{\link{FitEF}}.}

\item{lower_bound}{Lower limit of null neighborhood.}

\item{upper_bound}{Upper limit of null neighborhood.}
}
\value{
Estimated mean and variance of the null component, and the proportion
  of observations arising from the null component.
}
\description{
Estimates parameters of the empirical null distribution by central matching
within a "null neighborhood". Observations in the null neighborhood are
assumed to arise from the null component of the mixture distribution.
}
