% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFDR.R
\name{FDR}
\alias{FDR}
\title{Local False Discovery Rate}
\usage{
FDR(z, f = NULL, n_bins = NULL, null_mean = 0, null_var = 1, null_prop = 1)
}
\arguments{
\item{z}{Observations.}

\item{f}{Mixture density function (required).}

\item{n_bins}{Bins.}

\item{null_mean}{Mean of empirical null.}

\item{null_var}{Variance of empirical null.}

\item{null_prop}{Null proportion.}
}
\value{
List including two functions, lfdr for calculating the local false
  discovery rate, and tFDR for calculating the tail false discovery rate.
}
\description{
Estimates the local false discovery rate from a collection of z scores.
}
