% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{Fit.EF}
\alias{Fit.EF}
\title{Fit Exponential Family Density}
\usage{
Fit.EF(z, b = NULL, d = 2, method = "poly")
}
\arguments{
\item{z}{Observations.}

\item{b}{Bins for discretizing}

\item{d}{Polynomial degree.}

\item{method}{Either "ns" for natural splines, or "poly" for orthogonal
polynomials. Default is "poly".}
}
\value{
A list containing a function \code{f} for evaluating the fitted
  density, and the estimated regression coefficients \code{beta}.
}
\description{
Fit an exponential family distribution of degree \code{d} to a 
sample of observations \code{z}.
}
