% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{FitEF}
\alias{FitEF}
\title{Fit Exponential Family Density}
\usage{
FitEF(z, n_bins = NULL, degree = 2, method = "poly")
}
\arguments{
\item{z}{Observations.}

\item{n_bins}{Bins for discretizing.}

\item{degree}{Polynomial degree.}

\item{method}{Either "ns" for natural splines, or "poly" for orthogonal
polynomials. Default is "poly".}
}
\value{
A list containing a function \code{f} for evaluating the fitted
  density, and the estimated regression coefficients \code{beta}.
}
\description{
Fit an exponential family distribution of specified degree to a
sample of observations.
}
