% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMixture.R
\name{fitNsDens}
\alias{fitNsDens}
\title{Fit Natural Spline Basis Exponential Family}
\usage{
fitNsDens(z, k, l)
}
\arguments{
\item{z}{Normalized statistics.}

\item{k}{Bins.}

\item{l}{Polynomial degree.}
}
\description{
Fit Natural Spline Basis Exponential Family
}
