% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMixture.R
\name{fitPolyDens}
\alias{fitPolyDens}
\title{Fit Polynomial Basis Exponential Family}
\usage{
fitPolyDens(z, k, l)
}
\arguments{
\item{z}{Normalized statistics.}

\item{k}{Bins.}

\item{l}{Polynomial degree.}
}
\description{
Fit Polynomial Basis Exponential Family
}
