% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncNorm.R
\name{FitTN}
\alias{FitTN}
\title{Estimate Mean and Scale of Truncated Normal}
\usage{
FitTN(z, lower_bound, upper_bound, max_it = 25, eps = 1e-08, report = FALSE)
}
\arguments{
\item{z}{Observations.}

\item{lower_bound}{Lower truncation limit.}

\item{upper_bound}{Upper truncation limit.}

\item{max_it}{Maximum iterations.}

\item{eps}{Tolerance.}

\item{report}{Report fitting progress?}
}
\value{
Vector containing the estimated location and scale of the truncated
  normal distribution.
}
\description{
Estimates the location and scale of a truncated normal distribution.
}
