% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncNorm.R
\name{fitTruncNorm}
\alias{fitTruncNorm}
\title{Estimate Meand and Scale of Truncated Normal}
\usage{
fitTruncNorm(z, a, b, maxit = 25, trace = F)
}
\arguments{
\item{z}{Data.}

\item{a}{Lower truncation limit.}

\item{b}{Upper truncation limit.}

\item{maxit}{Maximum iterations.}

\item{trace}{Trace log likelihood.}
}
\description{
Estimate Meand and Scale of Truncated Normal
}
