% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenMixture.R
\name{genMixZ}
\alias{genMixZ}
\title{Generate Z's from Two-Component Normal-Mixutre}
\usage{
genMixZ(pi0, mu1, sigma1 = 1, n, symm = F)
}
\arguments{
\item{pi0}{Proportion of Z's from the null density.}

\item{mu1}{Location of alternative density.}

\item{sigma1}{Spread of alternative density.}

\item{n}{Number of statistics to generate.}

\item{symm}{If TRUE, the alternative is symmetric about zero.}
}
\description{
Generate Z's from Two-Component Normal-Mixutre
}
