% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncNorm.R
\name{H}
\alias{H}
\title{Truncated Normal Renormalizing Constant}
\usage{
H(loc_param, scale_param, lower_bound, upper_bound, order = 0, direction)
}
\arguments{
\item{loc_param}{Location parameter.}

\item{scale_param}{Scale parameter.}

\item{lower_bound}{Lower truncation limit.}

\item{upper_bound}{Upper truncation limit.}

\item{order}{Order.}

\item{direction}{Direction.}
}
\description{
Calculates the renormalizing constant of the truncated
normal distribution, and its derivatives.
}
