% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncNorm.R
\name{InfoTN}
\alias{InfoTN}
\title{Truncated Normal Observed Information}
\usage{
InfoTN(z, loc_param, log_scale, lower_bound, upper_bound)
}
\arguments{
\item{z}{Observations.}

\item{loc_param}{Location parameter.}

\item{log_scale}{Log of scale parameter.}

\item{lower_bound}{Lower truncation limit.}

\item{upper_bound}{Upper truncation limit.}
}
\description{
Evaluates the observed information of the sample log likelihood.
}
