% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncNorm.R
\name{initEst}
\alias{initEst}
\title{Initial Estimates of Mean and Scale for Truncated Normal}
\usage{
initEst(z, a, b, maxit = 5)
}
\arguments{
\item{z}{Data}

\item{a}{Lower truncation limit}

\item{b}{Upper truncation limit}

\item{maxit}{Maximum iterations}
}
\description{
Initial Estimates of Mean and Scale for Truncated Normal
}
