% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help.R
\docType{package}
\name{LFDR-help}
\alias{LFDR-help}
\alias{LFDR-help-package}
\title{LFDR: Local False Discovery Rate}
\description{
Utilities for local false discovery rate analysis. \code{\link{Fit.EF}} fits
an exponential family distribution of specified degree. \code{\link{eNull}}
estimates the empirical null distribution. \code{\link{FDR}} calculates the
local and tail false discovery rates.
}
\author{
Zachary R. McCaw
}
