% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFDR.R
\name{LFDR}
\alias{LFDR}
\title{Local False Disocery Rate}
\usage{
LFDR(z, f, k, m = 0, v = 1, p0 = 1)
}
\arguments{
\item{z}{Observed z scores}

\item{f}{Mixture density function}

\item{k}{Number of bins}

\item{m}{Mean of empirical null}

\item{v}{Variance of empirical null}

\item{p0}{Null proportion}
}
\value{
List including 1. a data frame, 2. a function for calculating the
  local FDR, and 3. a function for calculating the tail area FDR. The data
  frame includes the observed z scores, the discretized interval into which
  the z score fell, the midpoint of the interval, and the estimated local FDR
  for that interval.
}
\description{
Calculate the local false discovery rate for a set of z scores. In addition, returns 
funcations for calculating the local and tail are FDRs.
}
