% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotDens}
\alias{plotDens}
\title{Plot Estimated Densities}
\usage{
plotDens(z, f.mix, f.en, k = 120)
}
\arguments{
\item{z}{Observed z scores}

\item{f.mix}{Mixture density, optional}

\item{f.en}{Empirical null density, optional}

\item{k}{Bins for histogram}
}
\description{
Plot Estimated Densities
}
