% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotFDR}
\alias{plotFDR}
\title{Plot False Discovery Rates}
\usage{
plotFDR(z, lfdr, tFDR, k = 120)
}
\arguments{
\item{z}{Observed z scores.}

\item{lfdr}{Local fdr function}

\item{tFDR}{Tail fdr function}

\item{k}{Bins for histogram}
}
\description{
Plot False Discovery Rates
}
