% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncNorm.R
\name{QTN}
\alias{QTN}
\title{Truncated Normal Log Likelihood}
\usage{
QTN(z, loc_param, log_scale, lower_bound, upper_bound)
}
\arguments{
\item{z}{Observations.}

\item{loc_param}{Location parameter.}

\item{log_scale}{Log of scale parameter.}

\item{lower_bound}{Lower truncation limit.}

\item{upper_bound}{Upper truncation limit.}
}
\description{
Evaluates the sample log likelihood of the truncated
normal distribution.
}
