% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{RNM}
\alias{RNM}
\title{Simulate from Normal Mixture Model}
\usage{
RNM(n_obs, n_comp = 1, pi = NULL, loc_params = 0, scale_vars = 1)
}
\arguments{
\item{n_obs}{Number of observations.}

\item{n_comp}{Number of mixture components. Defaults to 1.}

\item{pi}{Mixture proportions. If omitted, components are assumed
equi-probable.}

\item{loc_params}{Vector of means, one for each component.}

\item{scale_vars}{Vector of variances, one for each component.}
}
\value{
Numeric matrix containing the mixture component and the observation.
}
\description{
Draws realizations from a normal mixture model. First, the cluster
membership is drawn from a multinomial distribution, with mixture proportions
specified by \code{pi}. Conditional on cluster membership, the observation is
drawn from a normal distribution, with cluster-specific mean and
variance. The cluster means are provided using \code{loc_params}, and the
cluster variances are provided using \code{scale_vars}.
}
\examples{
\dontrun{
# Two component mixture
Y <- RNM(n_obs = 1e3, n_comp = 2, pi = c(0.95, 0.05),
         loc_params = c(0, 2), scale_vars = c(1, 1))
}
}
