% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{RTN}
\alias{RTN}
\title{Simulate from Truncated Normal Distribution}
\usage{
RTN(n_obs, lower_bound, upper_bound, loc_param = 0, scale_param = 1)
}
\arguments{
\item{n_obs}{Number of observations.}

\item{lower_bound}{Lower truncation limit.}

\item{upper_bound}{Upper truncation limit.}

\item{loc_param}{Location parameter.}

\item{scale_param}{Scale parameter.}
}
\value{
Vector of realizations from the truncated normal distribution.
}
\description{
Draws realizations from a truncated normal distribution.
}
