% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{Sample.EF}
\alias{Sample.EF}
\title{Sample Fitted Exponential Family}
\usage{
Sample.EF(z, f, m = NULL, v = NULL, n = 1000, parallel = F,
  maxit = 25)
}
\arguments{
\item{z}{Observations.}

\item{f}{Density to sample.}

\item{m}{Proposal mean.}

\item{v}{Proposal variance.}

\item{n}{Target number of sample to obtain.}

\item{parallel}{Run in parallel? Must register parallel backend first.}

\item{maxit}{Maximum number of accept-reject iterations.}
}
\value{
Vector of realizations from the mixture density.
}
\description{
Sampling via the accept-reject algorithm. The proposal distribution
is taken as normal, with the option to specify the mean and variance.
}
