% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{SampleEF}
\alias{SampleEF}
\title{Sample Fitted Exponential Family}
\usage{
SampleEF(
  z,
  f,
  prop_mean = NULL,
  prop_var = NULL,
  n_samp = 1000,
  parallel = FALSE,
  max_it = 25
)
}
\arguments{
\item{z}{Observations.}

\item{f}{Density to sample.}

\item{prop_mean}{Proposal mean.}

\item{prop_var}{Proposal variance.}

\item{n_samp}{Target number of samples to obtain.}

\item{parallel}{Ignored; retained for backward compatibility.}

\item{max_it}{Maximum number of accept-reject iterations.}
}
\value{
Vector of realizations from the mixture density.
}
\description{
Sampling via the accept-reject algorithm. The proposal distribution
is taken as normal, with the option to specify the mean and variance.
}
