% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMixture.R
\name{sampleFit}
\alias{sampleFit}
\title{Sample Fitted Exponential Family}
\usage{
sampleFit(f, m = 0, s = 1, n, R, parallel = F, maxit = 25)
}
\arguments{
\item{f}{Density to sample.}

\item{m}{Proposal mean.}

\item{s}{Proposal standard deviation.}

\item{n}{Approximate number of sample to obtain.}

\item{R}{Range of data.}

\item{parallel}{Run in parallel?}

\item{maxit}{Maximum iterations of accept-reject.}
}
\description{
Sampling via the accept-reject algorithm. The proposal distribution
is taken as normal, with the option to specify the mean and variance.
}
