% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{Support}
\alias{Support}
\title{Support}
\usage{
Support(z, sig_digits = 1)
}
\arguments{
\item{z}{Observations.}

\item{sig_digits}{Significant digits.}
}
\value{
Numeric vector containing a lower and upper bound.
}
\description{
Return an interval containing the input data.
}
