% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncNorm.R
\name{UTN}
\alias{UTN}
\title{Truncated Normal Score Equations}
\usage{
UTN(z, loc_param, log_scale, lower_bound, upper_bound)
}
\arguments{
\item{z}{Observations.}

\item{loc_param}{Location parameter.}

\item{log_scale}{Log of scale parameter.}

\item{lower_bound}{Lower truncation limit.}

\item{upper_bound}{Upper truncation limit.}
}
\description{
Evaluates the gradient of the sample log likelihood.
}
