% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENull.R
\name{eNull}
\alias{eNull}
\title{Estimation of the Empirical Null}
\usage{
eNull(z, L = NULL, U = NULL, p = NULL, method = "ML", f = NULL)
}
\arguments{
\item{z}{Observations.}

\item{L}{Lower limit of null neighborhood.}

\item{U}{Upper limit of null neighborhood.}

\item{p}{Central proportion of z scores belonging to null neighborhood.}

\item{method}{Either 1. CM central matching, or 2. ML maximum likelihood}

\item{f}{Estimated mixture density if \code{method="CM"}.}
}
\value{
List containing the estimated mean and variance of the null
  component, and the proportion of observations arising from the null
  component.
}
\description{
Estimates the parameters of the empirical null distribution using
observations within a "null neighborhood". Observations in the null
neighborhood are assumed to arise from the null component of the mixture
distribution. The null neighborhood is specified using either a lower and
upper bound, or the central proportion of z scores supposed to arise from the
null.
}
