% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricalNull.R
\name{eNull}
\alias{eNull}
\title{Estimation of the Empirical Null}
\usage{
eNull(z, a, b, p, method = "ML", f)
}
\arguments{
\item{z}{Observed z scores}

\item{a}{Lower limit of null neighborhood}

\item{b}{Upper limit of null neighborhood}

\item{p}{Central proportion of z scores belonging to null neighborhood}

\item{method}{Either 1. CM central matching, or 2. ML maximum likelihood}

\item{f}{Estimated mixture density if \code{method="CM"}.}
}
\description{
Estimate mean \eqn{\mu}, variance \eqn{\sigma^2}, and null proportion
\eqn{\pi_{0}}. Specify either the absolutely limits of the null neighborhood,
or the central proportion of the observations to serve as the null
neighborhood.
}
