% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricalNull.R
\name{eNullCm}
\alias{eNullCm}
\title{Central-Matching Estimate of the Empirical Null}
\usage{
eNullCm(z, f, a, b)
}
\arguments{
\item{z}{Observed z scores}

\item{f}{Mixture density}

\item{a}{Lower limit of null neighborhood}

\item{b}{Upper limit of null neighborhood}
}
\description{
Estimate mean \eqn{\mu}, variance \eqn{\sigma^2}, and null proportion \eqn{\pi_{0}} 
by central matching.
}
