% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFDR.R
\name{FDR}
\alias{FDR}
\title{Local False Disocery Rate}
\usage{
FDR(z, f = NULL, b = NULL, m0 = 0, v0 = 1, p0 = 1)
}
\arguments{
\item{z}{Observations.}

\item{f}{Mixture density.}

\item{b}{Bins.}

\item{m0}{Mean of empirical null.}

\item{v0}{Variance of empirical null.}

\item{p0}{Null proportion.}
}
\value{
List including two functions, lfdr for calculating the local false
  discovery rate, and tFDR for calculating the tail false discovery rate.
}
\description{
Estimates the local false discovery rate from a collection of z scores.
}
