% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncNorm.R
\name{fitTN}
\alias{fitTN}
\title{Estimate Mean and Scale of Truncated Normal}
\usage{
fitTN(z, L, U, maxit = 25, eps = 1e-08, report = F)
}
\arguments{
\item{z}{Observations.}

\item{L}{Lower truncation limit.}

\item{U}{Upper truncation limit.}

\item{maxit}{Maximum iterations.}

\item{eps}{Tolerance.}

\item{report}{Report fitting progress?}
}
\value{
Vector containing the estimated location and scale of the truncated
  normal distribution.
}
\description{
Estimates the location and scale of a truncated normal distribution.
}
