% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncNorm.R
\name{H}
\alias{H}
\title{Truncated Normal Renormalizing Constant}
\usage{
H(m, s, L, U, k = 0, dirn)
}
\arguments{
\item{m}{Location parameter.}

\item{s}{Scale parameter.}

\item{L}{Lower truncation limit.}

\item{U}{Upper truncation limit.}

\item{k}{Order.}

\item{dirn}{Direction.}
}
\description{
Calculates the renormalizing constant of the truncated
normal distribution, and its derivatives.
}
