% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncNorm.R
\name{infoTN}
\alias{infoTN}
\title{Truncated Normal Observed Information}
\usage{
infoTN(z, m, t, L, U)
}
\arguments{
\item{z}{Observations.}

\item{m}{Location parameter.}

\item{t}{Log of scale parameter.}

\item{L}{Lower truncation limit.}

\item{U}{Upper truncation limit.}
}
\description{
Evaluates the observed information of the sample log likelihood.
}
