% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncNorm.R
\name{qTN}
\alias{qTN}
\title{Truncated Normal Log Likelihood}
\usage{
qTN(z, m, t, L, U)
}
\arguments{
\item{z}{Observations.}

\item{m}{Location parameter.}

\item{t}{Log of scale parameter.}

\item{L}{Lower truncation limit.}

\item{U}{Upper truncation limit.}
}
\description{
Evaluates the sample log likelihood of the truncated
normal distribution.
}
