% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rNM}
\alias{rNM}
\title{Simulated from Normal Mixture Model}
\usage{
rNM(n, k = 1, pi = NULL, m = 0, v = 1)
}
\arguments{
\item{n}{Observations.}

\item{k}{Number of mixture components. Defaults to 1.}

\item{pi}{Mixture proportions. If omitted, components are assumed
equi-probable.}

\item{m}{Vector of means, one for each component \eqn{k}.}

\item{v}{Vector of variances, one for each component \eqn{k}.}
}
\value{
Numeric matrix containing the mixture component and the observation.
}
\description{
Draws realizations from a normal mixture model. First, the cluster
membership is drawn from a multinomial distribution, with mixture proportions
specified by \code{pi}. Conditional on cluster membership, the observation is
drawn from a multivariate normal distribution, with cluster-specific mean and
variance. The cluster means are provided using \code{m}, and the cluster
covariance matrices are provided using \code{v}.
}
\examples{
\dontrun{
# Two component mixture
Y = rNM(n=1e3,k=2,pi=c(0.95,0.05),m=c(0,2),v=c(1,1));
}
}
