% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rTN}
\alias{rTN}
\title{Simulated from Truncated Normal Distribution}
\usage{
rTN(n, L, U, m = 0, s = 1)
}
\arguments{
\item{n}{Observations.}

\item{L}{Lower truncation limit.}

\item{U}{Upper truncation limit.}

\item{m}{Location parameter.}

\item{s}{Scale parameter.}
}
\value{
Vector of realizations from the truncated normal distribution.
}
\description{
Draws realizations from a truncated normal distribution.
}
